+++ test/eslint/helper.js
@@ -12,49 +12,6 @@ module.exports = {
 	concat: (...args) => args.join('\n'),
 
 	/**
-	 * @param {{valid: Array, invalid: Array}} test
-	 * @return {{valid: Array, invalid: Array}}
-	 */
-	extendToClassExpression: (test) => {
-		/**
-		 * @param {Array} testCases
-		 */
-		function extendToClassExpressionIfPossible(testCases) {
-			testCases.slice()
-				.forEach((testCase) => {
-					const extendedTestCase = {};
-
-					const classDeclarationRegExp = /^class (?:.*) {/;
-					const classDeclarationReplaceRegExp = /class ([^ ]*)/;
-					const classDeclarationReplaceString = '$1 = class';
-
-					if (testCase.code && classDeclarationRegExp.test(testCase.code)) {
-						extendedTestCase.code = testCase.code.replace(
-							classDeclarationReplaceRegExp,
-							classDeclarationReplaceString
-						);
-					}
-
-					if (testCase.output && classDeclarationRegExp.test(testCase.output)) {
-						extendedTestCase.output = testCase.output.replace(
-							classDeclarationReplaceRegExp,
-							classDeclarationReplaceString
-						);
-					}
-
-					if (Object.keys(extendedTestCase).length) {
-						testCases.push(Object.assign({}, testCase, extendedTestCase));
-					}
-				});
-		}
-
-		extendToClassExpressionIfPossible(test.valid);
-		extendToClassExpressionIfPossible(test.invalid);
-
-		return test;
-	},
-
-	/**
 	 * @param {string} text
 	 * @param {{valid: Array, invalid: Array}} test
 	 * @return {{valid: Array, invalid: Array}}
@@ -78,6 +35,77 @@ module.exports = {
 
 		prepend(test.valid);
 		prepend(test.invalid);
+
+		return test;
+	},
+
+	/**
+	 * @param {{valid: Array, invalid: Array}} test
+	 * @return {{valid: Array, invalid: Array}}
+	 */
+	extendClassDeclarations: (test) => {
+		/**
+		 * @param {Array} testCases
+		 */
+		function extend(testCases) {
+			testCases.slice()
+				.forEach((testCase) => {
+					const replacers = [
+						assignmentReplacer,
+						assignmentToMemberExpressionReplacer,
+						variableDeclarationReplacer
+					];
+
+					const declarationRegExp = /class (.*?) /g;
+					const extendedTestCases = Array.from(Array(replacers.length)).map(() => ({}));
+
+					let execResult;
+
+					// eslint-disable-next-line no-cond-assign
+					while (execResult = declarationRegExp.exec(testCase.code)) {
+						// eslint-disable-next-line no-loop-func
+						replacers.forEach((replacer, index) => {
+							extendedTestCases[index].code = (
+								replacer(extendedTestCases[index].code || testCase.code, execResult[1])
+							);
+						});
+					}
+
+					declarationRegExp.lastIndex = 0;
+
+					// eslint-disable-next-line no-cond-assign
+					while (execResult = declarationRegExp.exec(testCase.output)) {
+						// eslint-disable-next-line no-loop-func
+						replacers.forEach((replacer, index) => {
+							extendedTestCases[index].output = (
+								replacer(extendedTestCases[index].output || testCase.output, execResult[1])
+							);
+						});
+					}
+
+					extendedTestCases.forEach((extendedTestCase) => {
+						if (Object.keys(extendedTestCase).length) {
+							testCases.push(Object.assign({}, testCase, extendedTestCase));
+						}
+					});
+				});
+		}
+
+		function assignmentReplacer(code, className) {
+			return code.replace(new RegExp(`class ${className}`), `${className} = class`);
+		}
+
+		function assignmentToMemberExpressionReplacer(code, className) {
+			return assignmentReplacer(code, className)
+				.replace(new RegExp(className, 'g'), `ns.${className}`);
+		}
+
+		function variableDeclarationReplacer(code, className) {
+			return code.replace(new RegExp(`class ${className}`), `const ${className} = class`);
+		}
+
+		extend(test.valid);
+		extend(test.invalid);
 
 		return test;
 	}