+++ test/eslint/rules/lines-around-class.js
@@ -1,6 +1,6 @@
-const {errors, concat, extendToClassExpression} = require(`../helper`);
+const {errors, concat, extendClassDeclarations} = require(`../helper`);
 
-module.exports = extendToClassExpression({
+module.exports = extendClassDeclarations({
 	valid: [{
 		options: [{
 			before: 2,
@@ -74,11 +74,11 @@ module.exports = extendToClassExpression({
 			collisionPriority: 'after'
 		}],
 		code: concat(
-			`goog.provide('Klass');`,
+			`goog.provide('Klass1');`,
 			`goog.provide('Klass2');`,
 			``,
 			``,
-			`class Klass {}`,
+			`class Klass1 {}`,
 			``,
 			`class Klass2 {}`,
 			``,
@@ -91,11 +91,11 @@ module.exports = extendToClassExpression({
 			collisionPriority: 'before'
 		}],
 		code: concat(
-			`goog.provide('Klass');`,
+			`goog.provide('Klass1');`,
 			`goog.provide('Klass2');`,
 			``,
 			``,
-			`class Klass {}`,
+			`class Klass1 {}`,
 			``,
 			``,
 			`class Klass2 {}`,
@@ -109,14 +109,14 @@ module.exports = extendToClassExpression({
 			collisionPriority: 'after'
 		}],
 		code: concat(
-			`goog.provide('Klass');`,
+			`goog.provide('Klass1');`,
 			`goog.provide('Klass2');`,
 			``,
 			``,
 			`/**`,
 			` * @abstract`,
 			` */`,
-			`class Klass {}`,
+			`class Klass1 {}`,
 			``,
 			`/**`,
 			` * @abstract`,
@@ -132,14 +132,14 @@ module.exports = extendToClassExpression({
 			collisionPriority: 'before'
 		}],
 		code: concat(
-			`goog.provide('Klass');`,
+			`goog.provide('Klass1');`,
 			`goog.provide('Klass2');`,
 			``,
 			``,
 			`/**`,
 			` * @abstract`,
 			` */`,
-			`class Klass {}`,
+			`class Klass1 {}`,
 			``,
 			``,
 			`/**`,
@@ -259,8 +259,6 @@ module.exports = extendToClassExpression({
 			after: 2
 		}],
 		code: concat(
-			`goog.provide('Klass');`,
-			``,
 			`function func() {`,
 			`    // Comment`,
 			``,
@@ -270,9 +268,36 @@ module.exports = extendToClassExpression({
 			`}`
 		),
 		output: concat(
-			`goog.provide('Klass');`,
-			``,
 			`function func() {`,
+			`    // Comment`,
+			``,
+			``,
+			`    class Klass {}`,
+			``,
+			``,
+			`    Klass.CONST = 1;`,
+			`}`
+		),
+		errors: errors(
+			`Amount of newlines before class should be 2, but 1 given.`,
+			`Amount of newlines after class should be 2, but 1 given.`
+		)
+	}, {
+		options: [{
+			before: 2,
+			after: 2
+		}],
+		code: concat(
+			`for (;;) {`,
+			`    // Comment`,
+			``,
+			`    class Klass {}`,
+			``,
+			`    Klass.CONST = 1;`,
+			`}`
+		),
+		output: concat(
+			`for (;;) {`,
 			`    // Comment`,
 			``,
 			``,